+++ CareForMe/View/Extensions/UIStackView+componentView.swift
@@ -1,0 +1,23 @@
+//
+//  UIStackView+componentView.swift
+//  CareForMe
+//
+//  Created by Kenneth Dubroff on 7/11/21.
+//
+
+import UIKit
+
+extension UIStackView {
+    static func componentStack(elements: [UIView]) -> UIStackView {
+        let stack = UIStackView(arrangedSubviews: elements)
+        stack.translatesAutoresizingMaskIntoConstraints = false
+        stack.axis = .vertical
+        stack.distribution = .fill
+        stack.alignment = .fill
+        stack.spacing = 2
+        
+        let horizontalStack = UIStackView(arrangedSubviews: [stack])
+        horizontalStack.alignment = .top
+        return horizontalStack
+    }
+}
